[add] PlayerFactoryを追加
それに伴いPlayerの生成方法を変更

+++ Script/Object/Enemy/Base/BaseEnemy.cpp
@@ -58,8 +58,8 @@ void BaseEnemy::Move(){
 
 	float x, y;
 
-	x = cos(moveAngle) * moveSpeed;
-	y = sin(moveAngle) * moveSpeed;
+	x = (float)(cos(moveAngle) * moveSpeed);
+	y = (float)(sin(moveAngle) * moveSpeed);
 	vec2.Add(x, y);
 
 }
